backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# HTTP client and async support
httpx>=0.25.0
aiofiles>=23.2.0

# Configuration and data handling
pyyaml>=6.0.1
python-dotenv>=1.0.0
python-multipart>=0.0.6

# OpenAI integration
openai>=1.3.0
tiktoken>=0.5.0

# Vector databases
pinecone-client>=2.2.4
chromadb>=0.4.15

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.7
asyncpg>=0.28.0

# Redis cache
redis>=5.0.0
aioredis>=2.0.1

# Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Monitoring and observability
prometheus-client>=0.17.1
structlog>=23.1.0

# Document processing
pypdf2>=3.0.1
python-docx>=0.8.11
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Text processing
nltk>=3.8.1
spacy>=3.7.0
sentence-transformers>=2.2.2

# Utilities
requests>=2.31.0
tenacity>=8.2.3
validators>=0.22.0
python-slugify>=8.0.1
